fix: move findAllByMentorId from mentor to lessons
